Clarify behavior when receiving a tracestate header without traceparent (#326)
As the processing model was written, it can be interpreted such that a
tracing system that receives a tracestate header without an
accompanying traceparent header is valid, and that it should append to
and propagate that tracestate. This commit clarifies that a tracestate
header received without a traceparent header is invalid, and a new
tracestate should be created.