Specify uniqueness of tracestate keys as the sender's responsibility (#477)
* Specify uniqueness of tracestate keys as the sender's responsibility

This partially resolves #446.

This adds a new rule to "Mutating the tracestate Field" that makes it
clear that a sender must not create duplicated tracestate keys. It is a
non-breaking change because the section "Combined Header Value" already
had called that out with "Only one entry per key is allowed". Basically
this adds that rule to the section it actually belongs into.

Having that as the responsibility of the sender seems to be consensus in

There seems to be no consensus yet whether an implementation should
clean up a tracestate value it receives by discarding duplicate keys
that it does not own. Thus, that aspect is not considered in this
change.

* remove redundant phrase about not adding your own key twice

This is already called out in the section
"Mutating the traceparent Field".

* call out that foreign duplicated keys MAY be deleted

fixes #446

* fix test case for duplicated tracestate keys

The test_tracestate_duplicated_keys was broken, it demanded to drop
the whole tracestate when a duplicate key was found. This does not match
what the specification mandates in that situation. Instead,
implementations MAY remove the duplicates from tracestate or even leave
them in as-is.

fixes #369