improve description of version field in traceresponse header
This fixes the confusing wording in the description of the version
field, which previously didn't clearly distinguish between a possible
internal representation (1 byte) and the format on the wire
(two ascii chars).

This change ports the new description of the version field in the
request header over to the response header and aligns both
descriptions. See https://github.com/w3c/trace-context/pull/511.